Simplify platform specific app setup
