+++ packages/studio-base/src/index.ts
@@ -19,44 +19,28 @@ export {
   default as DialogHostIdContext,
   useDialogHostId,
 } from "@foxglove/studio-base/context/DialogHostIdContext";
-export { default as App } from "./App";
+export { ActualApp } from "./ActualApp";
 export type { NetworkInterface, OsContext } from "./OsContext";
-export { default as ErrorBoundary } from "./components/ErrorBoundary";
-export { default as MultiProvider } from "./components/MultiProvider";
-export { default as AppConfigurationContext } from "./context/AppConfigurationContext";
 export type {
-  AppConfiguration,
+  IAppConfiguration,
   AppConfigurationValue,
   ChangeHandler,
 } from "./context/AppConfigurationContext";
-export { default as LayoutStorageContext, useLayoutStorage } from "./context/LayoutStorageContext";
 export type { Layout, LayoutID, ISO8601Timestamp, ILayoutStorage } from "./services/ILayoutStorage";
 export { migrateLayout } from "./services/ILayoutStorage";
-export { HoverValueProvider } from "./context/HoverValueContext";
-export { UserNodeStateProvider } from "./context/UserNodeStateContext";
-export { default as NativeAppMenuContext } from "./context/NativeAppMenuContext";
-export type { NativeAppMenu, NativeAppMenuEvent } from "./context/NativeAppMenuContext";
+export type { INativeAppMenu, NativeAppMenuEvent } from "./context/NativeAppMenuContext";
 export { default as NativeWindowContext } from "./context/NativeWindowContext";
-export type { NativeWindow } from "./context/NativeWindowContext";
+export type { INativeWindow } from "./context/NativeWindowContext";
 export type { IDataSourceFactory } from "./context/PlayerSelectionContext";
-export { default as ThemeProvider } from "./theme/ThemeProvider";
 export { default as installDevtoolsFormatters } from "./util/installDevtoolsFormatters";
 export { makeMockAppConfiguration } from "./util/makeMockAppConfiguration";
 export { default as overwriteFetch } from "./util/overwriteFetch";
 export { default as waitForFonts } from "./util/waitForFonts";
 export { default as UserProfileLocalStorageProvider } from "./providers/UserProfileLocalStorageProvider";
-export { default as StudioToastProvider } from "./components/StudioToastProvider";
-export { default as ExtensionLoaderContext } from "./context/ExtensionLoaderContext";
 export type { ExtensionLoader, ExtensionInfo } from "./context/ExtensionLoaderContext";
 export { default as LayoutManagerContext } from "./context/LayoutManagerContext";
 export { AppSetting } from "./AppSetting";
-export { useAppConfigurationValue } from "./hooks/useAppConfigurationValue";
-export type { PanelsState } from "./context/CurrentLayoutContext/actions";
-export { default as LayoutStorageDebuggingContext } from "./context/LayoutStorageDebuggingContext";
-export { default as CssBaseline } from "./components/CssBaseline";
-export { default as GlobalCss } from "./components/GlobalCss";
-export { default as CurrentUserContext } from "./context/CurrentUserContext";
-export type { User } from "./context/CurrentUserContext";
+export type { User, CurrentUser } from "./context/CurrentUserContext";
 export { default as DeviceCodeDialog } from "./components/DeviceCodeDialog";
 export { default as FoxgloveDataPlatformDataSourceFactory } from "./dataSources/FoxgloveDataPlatformDataSourceFactory";
 export { default as FoxgloveWebSocketDataSourceFactory } from "./dataSources/FoxgloveWebSocketDataSourceFactory";
@@ -71,4 +55,3 @@ export { default as VelodyneDataSourceFactory } from "./dataSources/VelodyneData
 export { default as McapLocalDataSourceFactory } from "./dataSources/McapLocalDataSourceFactory";
 export { default as McapRemoteDataSourceFactory } from "./dataSources/McapRemoteDataSourceFactory";
 export { default as SampleNuscenesDataSourceFactory } from "./dataSources/SampleNuscenesDataSourceFactory";
-export { default as ReadySignalContext, useReadySignal } from "./stories/ReadySignalContext";
