+++ web/src/hooks/useCurrentUser.ts
@@ -2,12 +2,12 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { PropsWithChildren, useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useAsync } from "react-use";
 
 import { useShallowMemo } from "@foxglove/hooks";
 import Logger from "@foxglove/log";
-import { useConsoleApi, CurrentUserContext, User } from "@foxglove/studio-base";
+import { User, ConsoleApi, CurrentUser } from "@foxglove/studio-base";
 
 const log = Logger.getLogger(__filename);
 
@@ -16,11 +16,7 @@ const log = Logger.getLogger(__filename);
  *
  * On mount, it loads the current user profile if there is a session.
  */
-export default function ConsoleApiCookieCurrentUserProvider(
-  props: PropsWithChildren<unknown>,
-): JSX.Element {
-  const api = useConsoleApi();
-
+export function useCurrentUser(api: ConsoleApi): CurrentUser {
   const [currentUser, setCurrentUser] = useState<User | undefined>();
 
   // initial load of the user profile
@@ -67,11 +63,5 @@ export default function ConsoleApiCookieCurrentUserProvider(
     window.location.href = `${process.env.FOXGLOVE_CONSOLE_URL}/signin?returnTo=${currentLocation}`;
   }, []);
 
-  const value = useShallowMemo({ currentUser, signIn, signOut });
-
-  if (isLoading) {
-    return <></>;
-  }
-
-  return <CurrentUserContext.Provider value={value}>{props.children}</CurrentUserContext.Provider>;
+  return useShallowMemo({ currentUser, signIn, signOut, isLoading });
 }
