+++ web/src/services/LocalStorageAppConfiguration.ts
@@ -2,9 +2,9 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { AppConfiguration, ChangeHandler, AppConfigurationValue } from "@foxglove/studio-base";
+import { IAppConfiguration, ChangeHandler, AppConfigurationValue } from "@foxglove/studio-base";
 
-export default class LocalStorageAppConfiguration implements AppConfiguration {
+export default class LocalStorageAppConfiguration implements IAppConfiguration {
   static KEY_PREFIX = "studio.app-configuration.";
 
   /** Default values for app configuration items which have never been set by a user */
